requirements.txt
networkx~=2.5
pyro-ppl
ray[tune]~=1.0.0
scikit-learn~=0.23.2
nflows~=0.13
torch~=1.7.0
matplotlib~=3.3.2
numpy~=1.19.3
seaborn~=0.11.0
typing~=3.7.4.3
sklearn~=0.0
scipy~=1.2.1
statsmodels~=0.12.1
pandas~=1.1.3
setuptools~=50.3.2
coloredlogs
hydra-core
hydra_colorlog
mlflow